Move chp_etrago dataset to new file
